requirements_security.txt
# 🔒 Dependencias de Seguridad - OWASP
# Actualizar requirements.txt con estas dependencias

# Dependencias existentes
Django==5.2.1
pymysql==1.1.0
Pillow==10.1.0
django-tailwind==3.8.0
cryptography==41.0.7
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0
pyotp==2.9.0
qrcode==7.4.2

# 🔒 NUEVAS DEPENDENCIAS DE SEGURIDAD

# Hash de contraseñas mejorado (Argon2)
argon2-cffi==23.1.0

# Sanitización HTML
bleach==6.1.0

# Análisis de seguridad
bandit==1.7.5

# Verificación de vulnerabilidades en dependencias
safety==3.0.1

# Rate limiting
django-ratelimit==4.1.0

# Content Security Policy
django-csp==3.8

# Validación de tipos MIME de archivos
python-magic==0.4.27

# Bcrypt para passwords
bcrypt==4.1.2

# CORS Headers
django-cors-headers==4.3.1

# Django extensions para seguridad
django-extensions==3.2.3

# SSL/TLS para producción
certifi==2023.11.17

